pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-comment-analysis"
version = "0.1.0"
description = "Production-oriented YouTube comment analysis project using the official YouTube Data API v3."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Codex" }]
dependencies = [
  "deep-translator>=1.11.4",
  "matplotlib>=3.8.0",
  "nltk>=3.9.0",
  "openpyxl>=3.1.2",
  "pandas>=2.2.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "requests>=2.31.0",
  "scikit-learn>=1.4.0",
  "streamlit>=1.44.0",
  "vaderSentiment>=3.3.2",
  "wordcloud>=1.9.3"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]

[project.scripts]
yt-comment-analysis = "src.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
